C Program to check whether the given integer is positive or negative
In the following program we are checking whether the input integer number is positive or negative. If the input number is greater than zero then its positive else it is a negative number. If the number is zero then it is neither positive nor negative. Same logic we have followed in the below C program.